+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Changed command suggestions / autocomplete to also work with aliases.
 
 ## [0.1.9-alpha] - 2021-08-05
 - Added events for when the console is enabled/disabled, opened/closed and focused/unfocused.